+++ packages/create-vite/src/index.ts
@@ -633,10 +633,10 @@ async function init() {
   if (isReactTemplate && eslint === undefined) {
     if (interactive) {
       const eslintResult = await prompts.select({
-        message: 'Use ESLint instead of Oxlint?',
+        message: 'Which linter to use?',
         options: [
-          { label: 'No (Oxlint)', value: false },
-          { label: 'Yes (ESLint)', value: true },
+          { label: 'Oxlint', value: false },
+          { label: 'ESLint', value: true },
         ],
       })
       if (prompts.isCancel(eslintResult)) return cancel()
@@ -852,15 +852,15 @@ function setupEslint(root: string, isTs: boolean) {
   // renovate: datasource=npm depName=@eslint/js
   const eslintJsVersion = '10.0.1'
   // renovate: datasource=npm depName=eslint
-  const eslintVersion = '10.5.0'
+  const eslintVersion = '10.6.0'
   // renovate: datasource=npm depName=eslint-plugin-react-hooks
   const eslintPluginReactHooksVersion = '7.1.1'
   // renovate: datasource=npm depName=eslint-plugin-react-refresh
   const eslintPluginReactRefreshVersion = '0.5.3'
   // renovate: datasource=npm depName=globals
-  const globalsVersion = '17.6.0'
+  const globalsVersion = '17.7.0'
   // renovate: datasource=npm depName=typescript-eslint
-  const typescriptEslintVersion = '8.61.0'
+  const typescriptEslintVersion = '8.62.0'
 
   const eslintConfigForTS = /* js */ `import js from '@eslint/js'
 import globals from 'globals'